Python Project Rangoli Pattern
This project started as a solution to a Hacker Rank problem.  I would like to extend it to allow the user to choose the repeating pattern, the edge pattern, the fill, the text and the fill for the text line (if different than the other fill).  The user can also upload images to use as a pattern.  This would be useful for creating tile patterns.  So I need to abstract the input and create a user interface and make an app out of it,